Implement InvitationManager

+++ RadHocComponents/src/radhoc/gamestates/GameStateManager.java
@@ -17,7 +17,7 @@ public interface GameStateManager {
 	 * @param gameID
 	 * @param playerStarts if true, this player has the first turn of the game (only applicable to some GameTypes)
 	 */
-	void createGameState(GameType gameType, String opponentName, int opponentID, int gameID, boolean playerStarts);
+	void createGameState(GameType gameType, String opponentName, long opponentID, long gameID, boolean playerStarts);
 	
 	/**
 	 * Sets the UpdateListener for this GameState.
